adding another page called Hobbies, and migrating App content to Home, making App just a router
