+++ my-portfolio/src/pages/Hobbies.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Hobbies() {
+  return (
+    <div>
+      <h1>Página Nueva</h1>
+    </div>
+  );
+}
